+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.32.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.32.0)
+_May 26, 2023_
+
 ### 🐞 Fixed
 - Fixed the text input cursor when a message is being edited
 - Fixed channel list view model always not using passed channel type for deletion